Open reset password page from forgot-password page on Reset btn click
