+++ src/components/burger-constructor/burger-constructor.tsx
@@ -6,8 +6,8 @@ import { OrderDetails } from '../order-details/order-details';
 import { Modal } from '../modal/modal';
 import { CategoryKey } from '../../enums/category-key.enum';
 import { useSelector } from 'react-redux';
-import { StoreInterface } from '../../services/reducers';
-import { addIngredientToConstructor, submitNewOrder } from '../../services/actions';
+import { StoreInterface } from '../../services/reducers/burger.reducer';
+import { addIngredientToConstructor, submitNewOrder } from '../../services/actions/burger.actions';
 import { useDrop } from 'react-dnd';
 import { DndIngredientType } from '../../utils/app.types';
 import { BurgerConstructorBetweenBuns } from './burger-constructor-between-buns/burger-constructor-between-buns';
