+++ src/interfaces/password-reset-response.interface.ts
@@ -1,0 +1,4 @@
+export interface PasswordResetResponseInterface {
+  success: boolean;
+  message: string;
+}
