+++ src/services/actions/auth.actions.interface.ts
@@ -1,0 +1,1 @@
+export interface AuthActionsInterface {}
